+++ tests/speedtest-server.cpp
@@ -8,8 +8,8 @@
 
 #include <CLI/Validators.hpp>
 #include <future>
-#include <quic.hpp>
-#include <quic/gnutls_crypto.hpp>
+#include <oxen/quic.hpp>
+#include <oxen/quic/gnutls_crypto.hpp>
 #include <thread>
 
 #include "utils.hpp"
